+++ RxExample/RxExample-iOSTests/TestScheduler+MarbleTests.swift
@@ -49,13 +49,13 @@ extension TestScheduler {
         let allExceptLast = timelines[0 ..< timelines.count - 1]
 
         return (allExceptLast.map { $0 + "|" } + [timelines.last!])
-            .filter { $0.characters.count > 0 }
+            .filter { $0.count > 0 }
             .map { timeline -> [Recorded<Event<T>>] in
                 let segments = timeline.components(separatedBy:"-")
                 let (time: _, events: events) = segments.reduce((time: 0, events: [RecordedEvent]())) { state, event in
-                    let tickIncrement = event.characters.count + 1
+                    let tickIncrement = event.count + 1
 
-                    if event.characters.count == 0 {
+                    if event.count == 0 {
                         return (state.time + tickIncrement, state.events)
                     }