+++ Sources/AllTestz/main.swift
@@ -708,6 +708,7 @@ final class SingleTest_ : SingleTest, RxTestCase {
     ("test_map", SingleTest.test_map),
     ("test_flatMap", SingleTest.test_flatMap),
     ("test_flatMapMaybe", SingleTest.test_flatMapMaybe),
+    ("test_flatMapCompletable", SingleTest.test_flatMapCompletable),
     ("test_asMaybe", SingleTest.test_asMaybe),
     ("test_zip_tuple", SingleTest.test_zip_tuple),
     ("test_zip_resultSelector", SingleTest.test_zip_resultSelector),